Refactor backend to be more framework agnostic
